Project Xero is Halted
Further development is put on an indefinite hiatus.

The dataset contains inconsistent data and metrics,
so the project will be halted until a resolution is defined.

+++ misc/entities.sql
@@ -21,7 +21,7 @@ CREATE TABLE PriceData(
     HighPrice Decimal(10, 2) UNSIGNED,
     LowPrice DECIMAL(10, 2) UNSIGNED,
     AdjClosePrice DECIMAL(10, 2) UNSIGNED,
-    Volume INT UNSIGNED,
+    Volume BIGINT UNSIGNED,
     PRIMARY KEY(StockID, TradeDate),
     FOREIGN KEY (StockID) REFERENCES Stocks(StockID)
 )
